Divide files into classes
